tracker-project/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
tldextract==5.1.1
pydantic==2.5.3
snowflake-connector-python==3.6.0
python-dotenv==1.0.0